feat: Action System (#367)
This PR implements the action system proposed in #304.

# Architectural Overview

## ActionManager

Implements features surrounding the action registry.

Exposes an API surface for (among other things) registering actions and
associating hotkeys with actions.

It's important to note that actions each have a unique path.

## AppActions

Registers the standard Mupen64 actions along with their hotkeys (e.g.
`File > Load ROM...`).

Currently also implements all the callbacks for the actions (e.g.
`load_rom()` for `File > Load ROM...). The implementation part might be
refactored into separate modules in the future, but that is outside the
scope of this PR.

## ActionMenu

Implements a "managed" menu which mirrors the action "hierarchy" of the
actions contained within the action registry. The hierarchy is derived
from the segments of an action's path.

Exposes an API surface for adding a managed menu to a top-level window.
The menu is currently only being added to the main window, but usage can
be extended to auxiliary windows such as the Piano Roll in the future.

## HotkeyTracker

Implements a mechanism for tracking key presses in a window and
triggering actions based on the hotkey<->action associations specified
in the config (and established with `ActionManager::associate_hotkey`).

## Config

The config system has been changed to store a map of action paths to
their respective hotkeys (and initial hotkeys).

## ConfigDialog

The config dialog's related option item abstraction has been refactored
to fit the changed concept of a hotkey.

No functional changes have been made other than separating the
monolithic hotkey option group into multiple groups (as represented by
the action path segments).

---

In addition to this, the `Hotkey` and `ActionManager` APIs are exposed
to Lua.